requirements.txt
flask[async]
gunicorn[gthread]
transformers[torch]
pandas
python-okx
requests
git+https://github.com/amazon-science/chronos-forecasting.git